Remove legacy text command in favor of slash command
